refactor(mobile): name shared markdown renderer without iOS suffixes
